fixed animmap not being smooth if game time is too big

+++ CODE-mp/cgame/tr_types.h
@@ -111,7 +111,7 @@ typedef struct miniRefEntity_s
 	float				rotation;			// size 2 for RT_CYLINDER or number of verts in RT_ELECTRICITY
 
 	// misc
-	float		shaderTime;			// subtracted from refdef time to control effect start times
+	double		shaderTime;			// subtracted from refdef time to control effect start times
 	int			frame;				// also used as MODEL_BEAM's diameter
 
 } miniRefEntity_t;
@@ -146,7 +146,7 @@ typedef struct {
 	float				rotation;
 
 	// misc
-	float		shaderTime;			// subtracted from refdef time to control effect start times
+	double		shaderTime;			// subtracted from refdef time to control effect start times
 	int			frame;				// also used as MODEL_BEAM's diameter
 	//
 	//
